GRN code
GRN backend code

+++ src/main/java/com/vroom/dbmodel/orm/Podetail.java
@@ -81,7 +81,7 @@ public class Podetail  implements java.io.Serializable {
     }
 
     @Transient
-    public int getTotalAmount(){
+    public double getTotalAmount(){
         return this.quantity * this.product.getProductcost();
     }
